Remove unnecessary casts
Clippy emits a bunch of warnings of form:

 warning: casting to the same type is unnecessary ...

As suggested, remove the unnecessary casts.

+++ bitcoin/src/pow.rs
@@ -198,7 +198,7 @@ impl Target {
             size += 1;
         }
 
-        CompactTarget(compact | (size << 24) as u32)
+        CompactTarget(compact | (size << 24))
     }
 
     /// Returns true if block hash is less than or equal to this [`Target`].
